Model/requirements.txt
fastapi
presidio_analyzer
presidio_anonymizer
presidio_image_redactor
presidio-analyzer[azure-ai-language]
pydantic
Pillow
spacy
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
urllib3
uvicorn
python-multipart
nest_asyncio
presidio-analyzer[azure-ai-language]